check_if_plugin_esists(): add failing test case
This is for a new behavior that I'll be using: instead of passing in the plugin
path; I want to pass in a plugin name. This is because the current
implementation simply checks if the argument it's given is a directory. It
doesn't check whether it is a subdirectory of $ASDF_HOME/plugins/