Significant documentation site improvements (#440)
# Summary

This contains fixes and significant improvements to the documentation site.

* fixes some routing issues due to erroneous docsify config
* fixed links and layout of 404 pg
* improved the language on the landing page as per #438 
* propose a NEW_README.md tackling concerns with #438  and #422 
* use docsify specific md to emphasise some text
* add docsify-themeable plugin to improve loading experience and future proof for customisation
* improve use of `asdf-vm` over `asdf`

## Other Information

Included in this PR is a file `docs/NEW_README.md` which is my proposed README for this repo. The intention was to have all documentation in a single place, the site, and simplify the readme bringing the focus to explaining the tool rather than instructing users on how to use it.

I will create a separate PR for the README, but in the meantime the landing page composes this new readme with the ballad.