zsh-native asdf completion setup
Add a zsh-native completion system function for `asdf`.

I too often hit snags with bash completions in zsh, so I bias strongly
towards zsh-native completions.

This needs to end up in `$fpath` with a filename of `_asdf` but it
seemed unfriendly to use that name in this repo, so call it `_asdf.zsh`
and document a symlink.  Mark the file as executable to be compatible
with one convention of "personal functions should only autoload if
executable" (but that's probably so historic that very few people still
use it; I'm one of them, though).