Ignore shim directory for executable lookups
Previously, the shims directory was only excluded for system versions,
which meant that `asdf exec` would fall back to a shim if the install
didn't have the desired binary. If the shims are up to date this isn't a
problem since we check the metadata, but it causes issues if those are
out of sync for whatever reason.

I've checked all the uses of this function and I'm not seeing anywhere
we should need the old behavior.