feat(latest): adds the flag --all to the latest command
As discussed in #1036, this patch adds a `--all` flag to the latest
command which will show the list of plugins, their latest version and if
this one is installed or not.

This patch basically reuses the logic used in the list command and the
latest command. The main different is the removal of the query
paramenter as different plugins might need different queries.