fix: more robust SHELL checks
While debugging an unrelated issue, I found that my `$SHELL --version`
output was returning as zsh but I actually run `fish` via nix. Instead
of using `$SHELL`, which isn't accurate in some cases, I've updated this
to attempt to use the various `*_VERSION` environment variables that are
available in the shells before failing back to the previous `$SHELL`
check. Realistically, `$SHELL --version` should be replaced with
something else but this should improve the current situation with
misidentifying shells in use.