versions.bash: Provide output in the format .tool-versions expects
I frequently find myself wanting to query the latest version of a thing,
and subsequently pinning it in my current repository.

At the moment, i have to manually munge the output, which is just the
version string, to be prefixed by the plugin name.

Previously:

```
$ asdf latest ruby 2
2.7.8
$ asdf latest ruby 2 >> .tool-versions
$ # ... pain / manual editing / argh
$ asdf install
ruby 2.7.8 installed
```

After this change:

```
$ asdf latest ruby 2
ruby 2.7.8
$ asdf latest ruby 2 >> .tool-versions
$ # no pain!
$ asdf install
ruby 2.7.8 installed
```