fix!: Be aware of `ForcePrepend`
Previously, the value of `ForcePrepend` was not used anywhere and was not behaving as specified.
With this commit, the behavior will be as specified.

BREAKING CHANGE:
So far, `ASDF_FORCE_PREPEND` environment variable setting has been ignored, but this commit may change the behavior depending on the setting.

If it is not set, macOS behavior will not change, but Linux behavior will change.
If you want to get the same behavior as before, it is recommended to set the relevant environment variable.

Referenced Documents:
- https://github.com/asdf-vm/asdf/blob/932ac468b7c24c2adef90a293a1f7280a0074cc4/docs/manage/configuration.md#asdf_force_prepend

Background:
@andrecloutier told about this issue me in a comment on another PullRequest. Thank you!
- https://github.com/asdf-vm/asdf/pull/2011#discussion_r1990316326